Rename sketcherBasicSMILES.h to sketcherBasicSMILES.h
