+++ test/coordgenBasicSMILES.h
